Add ability to specify runner as a string with args for all runners
This allows the user to specify the runner they want to use as a
string for all test runners. They can then pass in the arguments they
wish to that runner - for example setting the reporter for xctool. It
will also replace the test runner command with the correct path for
xctool and xcodebuild.